+++ src/MetroTRK/mem_TRK.h
@@ -3,8 +3,8 @@
 
 #include <Runtime/platform.h>
 
-SECTION_INIT void* TRK_memcpy(void* dst, const void* src, unsigned long n);
-SECTION_INIT void* TRK_memset(void* dst, int val, unsigned long n);
+SECTION_INIT void* TRK_memcpy(void* dst, const void* src, size_t n);
+SECTION_INIT void* TRK_memset(void* dst, int val, size_t n);
 void TRK_fill_mem(void* dest, int val, size_t count);
 
 #endif
